requirements.txt

configparser

Markdown==2.6.6

PyMySQL==0.7.9

tornado==4.4.1

savReaderWriter==3.4.2

statsmodels

whoosh

jieba

sqlalchemy

pymongo

xlrd

pandas

patsy

scipy==0.18.1

python-memcached==1.59

redis==2.10.6

werkzeug

